[refactor] Migrate CartIcon with quantity display to separated UI component
