+++ src/layouts/Navbar/Navbar.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
-import { useSelector } from "react-redux";
 import classes from "./Navbar.module.scss";
+
+import CartIcon from "../../components/UI/CartIcon/CartIcon";
 
 import Logo from "../../components/UI/Logo/Logo";
 import hamburger from "../../assets/icon/hamburger.svg";
-import cartIcon from "../../assets/icon/cartIcon.svg";
 import profile from "../../assets/icon/profile.svg";
 import close from "../../assets/icon/close.svg";
 
@@ -12,13 +12,6 @@ import { Link } from "react-router-dom";
 
 const Navbar = () => {
   const [isMenuOpened, setIsMenuOpened] = useState(false);
-  const { cart } = useSelector((state: any) => state.cart);
-  console.log(cart.reduce((acc: number, item: any) => acc + item.quantity, 0));
-
-  const cartQuantity = cart.reduce(
-    (acc: number, item: any) => acc + item.quantity,
-    0
-  );
 
   const toggleMenuHandler = () => {
     setIsMenuOpened((prevState) => !prevState);
@@ -37,14 +30,7 @@ const Navbar = () => {
           <Link to="#">
             <img src={profile} alt="profile button" />
           </Link>
-          <Link to="/cart">
-            <img src={cartIcon} alt="cart button" />
-            {cartQuantity > 0 && (
-              <div className={classes.quantity_counter}>
-                <span>{cartQuantity}</span>
-              </div>
-            )}
-          </Link>
+          <CartIcon />
         </div>
       </div>
       <div
